Restructure code related to libindy proofs
