--- libvcx/src/indyvc/proofs/mod.rs
@@ -1,2 +1,0 @@
-pub mod proof_request;
-pub mod proof_message;
