+++ libvcx/src/libindy/proofs/mod.rs
@@ -1,0 +1,4 @@
+pub mod prover;
+pub mod verifier;
+pub mod proof_request;
+mod proof_request_internal;
