+++ libvcx/src/libindy/proofs/prover/mod.rs
@@ -1,0 +1,2 @@
+pub mod prover;
+mod prover_internal;
