+++ libvcx/src/libindy/proofs/verifier/mod.rs
@@ -1,0 +1,2 @@
+pub mod verifier;
+mod verifier_internal;
